Misc rendering improvements: - Minor performance improvements by looking up cached models using computeIfAbsent rather than changing contains/get and put calls - Made it easier to interact with Model3D and adjust values for them - Fixed RenderData having different hashes based on position causing extra models to have to be cached when they are the same just positioned differently - Cleaned up some duplicate code related to having TERs render models - Improve how model data is handled for the QIO Redstone Adapter - Allow Fluid Tank Item renderer to use the normal Fluid Tank's cached models - Only mark QIO blocks as needing a model update if something about them changed - Fix a minor amount of clipping when rendering teleporters in specific orientations - Created a currently disabled system for rendering the fission control rod glow using baked models and custom model data rather than a TER to improve performance (currently disabled due to other rendering issues and transparency sorting)

+++ src/main/java/mekanism/client/render/RenderResizableCuboid.java
@@ -44,9 +44,9 @@ public class RenderResizableCuboid {
         // that may improve performance some, but definitely would reduce/remove the majority of remaining z-fighting that is going on
         //Shift it so that the min values are all greater than or equal to zero as the various drawing code
         // has some issues when it comes to handling negative numbers
-        float xShift = Mth.floor(cube.minX);
-        float yShift = Mth.floor(cube.minY);
-        float zShift = Mth.floor(cube.minZ);
+        int xShift = Mth.floor(cube.minX);
+        int yShift = Mth.floor(cube.minY);
+        int zShift = Mth.floor(cube.minZ);
         matrix.pushPose();
         matrix.translate(xShift, yShift, zShift);
         float minX = cube.minX - xShift;
